+++ lib/internal/bootstrap/node.js
@@ -123,7 +123,7 @@ process.emitWarning = emitWarning;
 const {
   nextTick,
   runNextTicks
-} = NativeModule.require('internal/process/next_tick').setup();
+} = NativeModule.require('internal/process/task_queues').setupTaskQueue();
 
 process.nextTick = nextTick;
 // Used to emulate a tick manually in the JS land.
@@ -201,7 +201,11 @@ if (!config.noBrowserGlobals) {
   defineOperation(global, 'clearTimeout', timers.clearTimeout);
   defineOperation(global, 'setInterval', timers.setInterval);
   defineOperation(global, 'setTimeout', timers.setTimeout);
-  setupQueueMicrotask();
+
+  const {
+    queueMicrotask
+  } = NativeModule.require('internal/process/task_queues');
+  defineOperation(global, 'queueMicrotask', queueMicrotask);
 
   // Non-standard extensions:
   defineOperation(global, 'clearImmediate', timers.clearImmediate);
@@ -393,17 +397,6 @@ function createGlobalConsole(consoleFromVM) {
     setConsoleExtensionInstaller(inspector.installConsoleExtensions);
   }
   return consoleFromNode;
-}
-
-function setupQueueMicrotask() {
-  const { queueMicrotask } =
-    NativeModule.require('internal/queue_microtask');
-  Object.defineProperty(global, 'queueMicrotask', {
-    value: queueMicrotask,
-    writable: true,
-    enumerable: true,
-    configurable: true,
-  });
 }
 
 function setupDOMException() {
