+++ lib/internal/modules/cjs/loader.js
@@ -870,7 +870,7 @@ Module.runMain = function() {
       return loader.import(pathToFileURL(process.argv[1]).pathname);
     })
     .catch((e) => {
-      internalBinding('util').triggerFatalException(e);
+      internalBinding('task_queue').triggerFatalException(e);
     });
   } else {
     Module._load(process.argv[1], null, true);
