+++ lib/internal/process/task_queues.js
@@ -6,7 +6,9 @@ const {
   tickInfo,
   // Used to run V8's micro task queue.
   runMicrotasks,
-  setTickCallback
+  setTickCallback,
+  enqueueMicrotask,
+  triggerFatalException
 } = internalBinding('task_queue');
 
 const {
@@ -27,7 +29,10 @@ const {
   emitDestroy,
   symbols: { async_id_symbol, trigger_async_id_symbol }
 } = require('internal/async_hooks');
-const { ERR_INVALID_CALLBACK } = require('internal/errors').codes;
+const {
+  ERR_INVALID_CALLBACK,
+  ERR_INVALID_ARG_TYPE
+} = require('internal/errors').codes;
 const FixedQueue = require('internal/fixed_queue');
 
 // *Must* match Environment::TickInfo::Fields in src/env.h.
@@ -130,15 +135,52 @@ function nextTick(callback) {
   queue.push(new TickObject(callback, args, getDefaultTriggerAsyncId()));
 }
 
-// TODO(joyeecheung): make this a factory class so that node.js can
-// control the side effects caused by the initializers.
-exports.setup = function() {
-  // Sets the per-isolate promise rejection callback
-  listenForRejections();
-  // Sets the callback to be run in every tick.
-  setTickCallback(processTicksAndRejections);
-  return {
-    nextTick,
-    runNextTicks
-  };
+let AsyncResource;
+function createMicrotaskResource() {
+  // Lazy load the async_hooks module
+  if (!AsyncResource) {
+    AsyncResource = require('async_hooks').AsyncResource;
+  }
+  return new AsyncResource('Microtask', {
+    triggerAsyncId: getDefaultTriggerAsyncId(),
+    requireManualDestroy: true,
+  });
+}
+
+function queueMicrotask(callback) {
+  if (typeof callback !== 'function') {
+    throw new ERR_INVALID_ARG_TYPE('callback', 'function', callback);
+  }
+
+  const asyncResource = createMicrotaskResource();
+
+  enqueueMicrotask(() => {
+    asyncResource.runInAsyncScope(() => {
+      try {
+        callback();
+      } catch (error) {
+        // TODO(devsnek) remove this if
+        // https://bugs.chromium.org/p/v8/issues/detail?id=8326
+        // is resolved such that V8 triggers the fatal exception
+        // handler for microtasks
+        triggerFatalException(error);
+      } finally {
+        asyncResource.emitDestroy();
+      }
+    });
+  });
+}
+
+module.exports = {
+  setupTaskQueue() {
+    // Sets the per-isolate promise rejection callback
+    listenForRejections();
+    // Sets the callback to be run in every tick.
+    setTickCallback(processTicksAndRejections);
+    return {
+      nextTick,
+      runNextTicks
+    };
+  },
+  queueMicrotask
 };
