+++ node.gyp
@@ -160,7 +160,6 @@
       'lib/internal/process/esm_loader.js',
       'lib/internal/process/execution.js',
       'lib/internal/process/main_thread_only.js',
-      'lib/internal/process/next_tick.js',
       'lib/internal/process/per_thread.js',
       'lib/internal/process/policy.js',
       'lib/internal/process/promises.js',
@@ -168,8 +167,8 @@
       'lib/internal/process/warning.js',
       'lib/internal/process/worker_thread_only.js',
       'lib/internal/process/report.js',
+      'lib/internal/process/task_queues.js',
       'lib/internal/querystring.js',
-      'lib/internal/queue_microtask.js',
       'lib/internal/readline.js',
       'lib/internal/repl.js',
       'lib/internal/repl/await.js',
