+++ test/message/events_unhandled_error_nexttick.out
@@ -13,7 +13,7 @@ Error
     at internal/main/run_main_module.js:*:*
 Emitted 'error' event at:
     at process.nextTick (*events_unhandled_error_nexttick.js:*:*)
-    at processTicksAndRejections (internal/process/next_tick.js:*:*)
-    at process.runNextTicks [as _tickCallback] (internal/process/next_tick.js:*:*)
+    at processTicksAndRejections (internal/process/task_queues.js:*:*)
+    at process.runNextTicks [as _tickCallback] (internal/process/task_queues.js:*:*)
     at Function.Module.runMain (internal/modules/cjs/loader.js:*:*)
     at internal/main/run_main_module.js:*:*
